feat: show empty cart view if removed last product from cart
